24 Linked List
Her I learned how to swap nodes the main thing here was the usage of a dummy node which represents or is needed for all cases where there is more than one node. My misconceptions was that i did not update the head node correctly and the dummy pointer is now he used before the head node and plus point here is that I can use it as better way to jump to the next as the pointer is to the third line.